A 2d, top-down, Bullet-hell game.
This is a game I've made for a project at my school. It was made in GameMaker Studio 2. As seen in the credits of the game most of the more advanced code was made by Microsoft Copilot. This is due to me never having coded a gamed before. Through this project I've learnt alot and my next game will probably take less time to develop when the time to make it comes.